+++ src/components/base/Component.ts
@@ -1,0 +1,1 @@
+export default class Component {}
