+++ src/stores/index.ts
@@ -1,4 +1,5 @@
-import { MovieResult } from "../../types/movie";
+import { MovieResult } from "../api/types/movie/response";
+
 export const store = {
   page: 1,
   totalPages: 1,
